bots/astra/gateway/pyproject.toml
[tool.poetry]
name = "astra-gateway"
version = "0.1.0"
description = "Astra gateway — GitHub webhook receiver and Cloud Tasks dispatcher"
authors = ["Fernando Correia"]
packages = [{include = "astra_gateway", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.135"
uvicorn = {version = "^0.44", extras = ["standard"]}
httpx = "^0.28"
PyJWT = {version = "^2.12", extras = ["crypto"]}
google-cloud-tasks = "^2.22"
google-cloud-run = "^0.16"
google-auth = "^2.49"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
pytest-asyncio = "^1.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
